Move "test" directory to "pkg/testutils"
Having top level directories called "test" and "tests" is
rather confusing. For even more confusion the go files in
the "test" directory declare they are in a go package called
"rest". Rename the "test" directory to "pkg/testutils" and
the go package to "testutils".
