+++ pkg/testutils/matchers.go
@@ -17,17 +17,19 @@
  *
  */
 
-package rest
+package testutils
 
 import (
 	"encoding/json"
 	"fmt"
+	"reflect"
+
 	"github.com/ghodss/yaml"
 	"github.com/onsi/gomega/types"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/client-go/rest"
+
 	rest2 "kubevirt.io/kubevirt/pkg/rest"
-	"reflect"
 )
 
 func RepresentMimeType(expected interface{}) types.GomegaMatcher {
